Started re-work according to last specification version

+++ src/Lykke.Service.Zcash.Api.Core/Domain/Insight/Utxo.cs
@@ -9,7 +9,7 @@ namespace Lykke.Service.Zcash.Api.Core.Domain.Insight
         public uint Vout { get; set; }
         public string ScriptPubKey { get; set; }
         public decimal Amount { get; set; }
-        public BigInteger Satoshis { get; set; }
+        public ulong Satoshis { get; set; } // UInt64 is enough for 2_100_000_000_000_000 of zatoshis
         public ulong? Confirmations { get; set; }
         public ulong? Height { get; set; }
     }
